requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.4
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
gunicorn==20.1.0
numpy==1.24.1
pandas==1.5.2
psycopg2==2.9.5
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.2.0
xlwt==1.3.0
